+++ packages/discussions/test/posts.test.ts
@@ -7,8 +7,9 @@ import {
   removePost,
   createReply,
   updatePostStatus,
+  exportPosts,
 } from "../src/posts";
-import * as req from "../src/request";
+import * as hubCommon from "@esri/hub-common";
 import {
   AclCategory,
   AclSubCategory,
@@ -16,6 +17,7 @@ import {
   IFetchPostParams,
   PostStatus,
   Role,
+  SearchPostsFormat,
   SharingAccess,
 } from "../src/types";
 
@@ -28,7 +30,7 @@ describe("posts", () => {
   } as unknown as IDiscussionsRequestOptions;
 
   beforeEach(() => {
-    requestSpy = spyOn(req, "request").and.returnValue(
+    requestSpy = spyOn(hubCommon, "discussionsApiRequest").and.returnValue(
       Promise.resolve(response)
     );
   });
@@ -137,6 +139,31 @@ describe("posts", () => {
               ...query,
               featureGeometry: JSON.stringify(query.featureGeometry),
             },
+          },
+          httpMethod: "GET",
+        });
+        done();
+      })
+      .catch(() => fail());
+  });
+
+  it("exports posts as a CSV string", (done) => {
+    const query = {
+      access: [SharingAccess.PUBLIC],
+      groups: ["foo"],
+    };
+
+    const options = { ...baseOpts, data: query };
+    exportPosts(options)
+      .then(() => {
+        expect(requestSpy.calls.count()).toEqual(1);
+        const [url, opts] = requestSpy.calls.argsFor(0);
+        expect(url).toEqual(`/posts`);
+        expect(opts).toEqual({
+          ...options,
+          data: {
+            ...options.data,
+            f: SearchPostsFormat.CSV,
           },
           httpMethod: "GET",
         });